requirements.txt
flask==3.1.2
requests==2.32.5
python-dotenv==1.1.1
huggingface-hub==0.34.4
transformers==4.56.0
torch==2.8.0
numpy==2.3.2
pandas==2.3.2
